requirements.txt
feedparser>=6.0
requests>=2.31
google-generativeai>=0.8
python-dotenv>=1.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
